+++ test/unit/lib/content_rendering_support_test.rb
@@ -18,7 +18,7 @@ class ContentRenderingSupportTest < ActiveSupport::TestCase
     assert c.respond_to? :handle_server_error_on_page
     assert c.respond_to? :handle_error_with_cms_page, true
 
-    assert(c.respond_to?(:logged_in?), "Should include Cms::Authentication::Controller methods")
+    assert(c.respond_to?(:logged_in?, true), "Should include Cms::Authentication::Controller methods")
   end
 
   class MyController < ActionController::Base
@@ -33,7 +33,7 @@ class ContentRenderingSupportTest < ActiveSupport::TestCase
     assert c.respond_to? :handle_server_error_on_page
     assert c.respond_to? :handle_error_with_cms_page, true
 
-    assert(!c.respond_to?(:logged_in?), "Should Not include Cms::Authentication::Controller methods")
+    assert(!c.respond_to?(:logged_in?, true), "Should Not include Cms::Authentication::Controller methods")
     #assert(c.respond_to?(:handle_server_error), "Should include Cms::ErrorHandling methods")
   end